refactor(toolbar): encapsulate components
