+++ src/components/index.ts
@@ -1,3 +1,4 @@
 import GitHubButton from "./GitHubButton"
+import LineDivider from "./LineDivider"
 
-export { GitHubButton }
+export { GitHubButton, LineDivider }
