+++ src/interfaces/editor/previewSettings.ts
@@ -1,6 +1,7 @@
 type PreviewSettings = {
   fontSize: number
   themeColor: string
+  paperColor: string
 }
 
 export type { PreviewSettings }
