+++ src/store/previewSettings.ts
@@ -1,17 +1,19 @@
 import { ref } from "vue"
 import { defineStore } from "pinia"
 import type { PreviewSettings } from "@/interfaces"
-import { lightThemeColorCandidates } from "@/constants"
 
 export const usePreviewSettingsStore = defineStore("previewSettings", () => {
   const previewSettings = ref<PreviewSettings>({
     fontSize: 13,
-    themeColor: lightThemeColorCandidates[0]
+    themeColor: "#000000",
+    paperColor: "#FFFFFF"
   })
 
   const changeFontSize = (fontSize: number) => (previewSettings.value.fontSize = fontSize)
 
   const changeThemeColor = (themeColor: string) => (previewSettings.value.themeColor = themeColor)
 
-  return { previewSettings, changeFontSize, changeThemeColor }
+  const changePaperColor = (paperColor: string) => (previewSettings.value.paperColor = paperColor)
+
+  return { previewSettings, changeFontSize, changeThemeColor, changePaperColor }
 })
