+++ src/views/Editor/Components/Toolbar/EditorTools/FontSize/index.vue
@@ -1,0 +1,50 @@
+<script setup lang="ts">
+import { NIcon, NSlider } from "naive-ui"
+import { TextFieldsOutlined as FontSizeIcon } from "@vicons/material"
+import { useEditorSettingsStore } from "@/store"
+import { formatTooltipWithPx } from "@/utils"
+
+const editorSettingsStore = useEditorSettingsStore()
+</script>
+
+<template>
+  <div class="block">
+    <div class="title-row">
+      <n-icon
+        size="20"
+        :depth="2"
+      >
+        <font-size-icon />
+      </n-icon>
+      Font Size
+    </div>
+    <n-slider
+      v-model:value="editorSettingsStore.editorSettings.fontSize"
+      :min="12"
+      :max="20"
+      :format-tooltip="formatTooltipWithPx"
+      class="slider"
+    />
+    <div class="scale-text">
+      <span>12px</span>
+      <span>16px</span>
+      <span>20px</span>
+    </div>
+  </div>
+</template>
+
+<style scoped lang="scss">
+@use "../../index.module.scss" as *;
+.block {
+  @include block;
+}
+.title-row {
+  @include title-row;
+}
+.scale-text {
+  @include scale-text;
+}
+.slider {
+  @include slider;
+}
+</style>
