+++ src/views/Editor/Components/Toolbar/index.module.scss
@@ -1,7 +1,7 @@
 @mixin container {
   display: flex;
   flex-direction: column;
-  gap: 8px;
+  gap: 12px;
   padding: 8px 16px;
 }
 @mixin block {
