+++ src/views/Test/ChildComponent/index.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts"></script>
+
+<template><div class="container"></div></template>
+
+<style scoped lang="scss">
+.container {
+  width: 100vw;
+  height: 100vh;
+  background-color: blue;
+}
+</style>
